IdleAnimation_Bukola
I added a Character folder & Animation folder. In the charcater folder I have a dumy_charcter that we can replace later with our offical charcter later. In the animation folder I have the Idle movement in the folder. I then created a Button in the Beginning Scenes called "IdleAnimation" that we'll use to call the animation(not sure about this part yet).